pyproject.toml
[tool.poetry]
name = "swiip"
version = "0.1.0"
description = "Swiips software chores"
authors = ["Kevin Lu"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
fastapi = "^0.94.1"
uvicorn = "^0.21.0"
PyGithub = "^1.58.0"
loguru = "^0.6.0"
nptyping = "^2.5.0"
pre-commit = "^3.2.0"
modal-client = "^0.48.1638"
requests = "^2.28.2"
types-requests = "^2.28.11.15"
tiktoken = "^0.3.2"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
pytest = "^7.2.2"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
check_untyped_defs = true
